requirements.txt
# requirements.txt

torch>=1.9.0
numpy>=1.20.0
PyYAML>=5.0
tqdm>=4.60.0